# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "main swagger APIs",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8888",
    "paths": {
        "/api/v1/pay/wechat/notify/:tid": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    ""
                ],
                "summary": "微信支付回调通知(URI包含交易ID)",
                "parameters": [
                    {
                        "description": "params description",
                        "name": "WechatPayNotify",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.WechatPayNotifyReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/sign_in": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    ""
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "params description",
                        "name": "UserSignIn",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserSignInReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.UserSignInRsp"
                        }
                    }
                }
            }
        },
        "/api/v1/sign_out": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    ""
                ],
                "summary": "用户退出登录",
                "parameters": [
                    {
                        "description": "params description",
                        "name": "UserSignOut",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserSignOutReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.UserSignOutRsp"
                        }
                    }
                }
            }
        },
        "/api/v1/sign_up": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    ""
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "params description",
                        "name": "UserSignUp",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserSignUpReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.UserSignUpRsp"
                        }
                    }
                }
            }
        },
        "/api/v1/user/:id": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    ""
                ],
                "summary": "根据ID查询用户",
                "parameters": [
                    {
                        "description": "params description",
                        "name": "GetUserById",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.GetUserByIdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GetUserByIdRsp"
                        }
                    }
                }
            }
        },
        "/api/v1/user/add": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    ""
                ],
                "summary": "添加用户",
                "parameters": [
                    {
                        "description": "params description",
                        "name": "AddUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.AddUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.AddUserRsp"
                        }
                    }
                }
            }
        },
        "/api/v1/user/delete": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    ""
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "description": "params description",
                        "name": "DeleteUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.DeleteUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.DeleteUserRsp"
                        }
                    }
                }
            }
        },
        "/api/v1/user/edit": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    ""
                ],
                "summary": "修改用户",
                "parameters": [
                    {
                        "description": "params description",
                        "name": "EditUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.EditUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.EditUserRsp"
                        }
                    }
                }
            }
        },
        "/api/v1/user/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    ""
                ],
                "summary": "用户列表",
                "parameters": [
                    {
                        "description": "params description",
                        "name": "GetUserList",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.GetUserListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GetUserListRsp"
                        }
                    }
                }
            }
        },
        "/api/v1/ws/market": {
            "get": {
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    ""
                ],
                "summary": "市场行情（websocket方式）",
                "parameters": [
                    {
                        "description": "params description",
                        "name": "WsMarketList",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "types.AddUserReq": {
            "type": "object"
        },
        "types.AddUserRsp": {
            "type": "object"
        },
        "types.DeleteUserReq": {
            "type": "object"
        },
        "types.DeleteUserRsp": {
            "type": "object"
        },
        "types.EditUserReq": {
            "type": "object"
        },
        "types.EditUserRsp": {
            "type": "object"
        },
        "types.GetUserByIdReq": {
            "type": "object"
        },
        "types.GetUserByIdRsp": {
            "type": "object"
        },
        "types.GetUserListReq": {
            "type": "object"
        },
        "types.GetUserListRsp": {
            "type": "object"
        },
        "types.UserSignInReq": {
            "type": "object"
        },
        "types.UserSignInRsp": {
            "type": "object"
        },
        "types.UserSignOutReq": {
            "type": "object"
        },
        "types.UserSignOutRsp": {
            "type": "object"
        },
        "types.UserSignUpReq": {
            "type": "object"
        },
        "types.UserSignUpRsp": {
            "type": "object"
        },
        "types.WechatPayNotifyReq": {
            "type": "object"
        }
    }
}